Improve nvim autocompletion
This requires a snippet engine to be installed even though I'm not
really using that functionality. This makes Tab more useful as it
doesn't just immediately complete the first found completion (which is
often just a word from the buffer), but requires a completion to be
selected first before committing to it.