Move undo and backup dirs
This creates a more standard format for undo and backup directories.
undodir being part of the dotfiles is a bit weird since they were
ignored anyway, and backups were actually not turned on so that didn't
do anything. That's now also turned on, and both have a standardized
temp path.